***

This PR implements the complete **v3.0 Mathematical Recipe** across the ML Core, Orchestrator, and Celery inference pipelines. It replaces the legacy Stage 2 binary thresholding with a continuous, fully bounded fractional Beta-Bernoulli Thompson Sampling engine.

By capturing continuous facial affect telemetry ($I_{bounded} \in [0, 1]$) and applying stimulus-locked robust aggregation, this update significantly improves the statistical efficiency and physiological accuracy of our adaptive experiment posteriors.

* **FACS-Anchored Calibration**: Updated the default $\alpha_{scale}$ from 5.0 to 6.0.
* **Soft-Saturation**: Introduced `compute_bounded_intensity()` which outputs strict $[0.0, 1.0]$ bounds via a $I_{bounded} = \tanh(\alpha \cdot \max(0, \Delta_{ratio}))$ transformation.
* **Backward Compatibility**: The original `compute_intensity()` with its $[0.0, 5.0]$ hard-clamp is preserved to prevent downstream dashboard breaks.

* **Per-Frame Extraction**: Integrated `MediaPipe` FaceMesh to calculate AU12 intensity on every video frame.
* **Calibration Lifecycle**: Added `record_stimulus_injection()` to transition the system from baseline accumulation ($B_{neutral}$) to active telemetry recording.
* **Payload Wiring**: `_au12_series` and `_stimulus_time` are now injected into the `assemble_segment()` handoff payload.

* **Stimulus-Locked Windowing**: Extracts active telemetry strictly from $[+0.5s, +5.0s]$ post-stimulus.
* **Robust Aggregation**: Computes the 90th percentile (P90) of the measurement window to reject single-frame noise while capturing sustained peaks.
* **Semantic Validity Gate**: Deterministically crushes the reward to $0.0$ if the context is a non-match ($r_t = P90 \times G_t$). Invalid observations (e.g., $<10$ frames) are gracefully skipped.

* **Continuous Updates**: Stripped the binary threshold (`reward >= 0.5`) in favor of continuous pseudo-counts. The Thompson Sampling update is now strictly:
    * $\alpha \leftarrow \alpha + r_t$
    * $\beta \leftarrow \beta + (1 - r_t)$
* **Persistence Wiring**: Wired the output of `compute_reward()` directly into the Celery `persist_metrics` task, including out-of-bounds guards and proper null-observation handling (missing data does not penalize the posterior).

* **Validation Suite**: Added a comprehensive 27-test mathematical recipe suite (`test_v3_math_recipe.py`).
* **Coverage**: Verified `tanh` bounds, raw ratio extraction, stimulus window timing, semantic gate logic, and fractional update convergence.
* **CI Status**: All 27 net-new tests passed. Full test colle...

+++ packages/ml_core/face_mesh.py
@@ -56,7 +56,7 @@ class FaceMeshProcessor:
             self.load_model()
 
         # §4.D.2 — MediaPipe expects RGB input
-        rgb_frame: npt.NDArray[np.uint8] = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
+        rgb_frame: npt.NDArray[np.uint8] = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB).astype(np.uint8)
         results: Any = self._face_mesh.process(rgb_frame)
 
         # §4.D contract — missing face returns None (null facial metrics)
